requirements.txt
# https://www.dl.dropboxusercontent.com/scl/fi/35w0i0hnsioq63x7eumdd/pyrogagan.zip?rlkey=mq4b56v0z7w7mgj3c0o3sqgo1&st=n7prfm1y&dl=0
pyrofork
psutil
aiofiles
pillow
# devgagantools - removed (old branding dependency)
tgcrypto
pyromod
opencv-python-headless
requests
motor
pytz
flask
aiohttp
telethon
aiojobs
werkzeug==2.2.2
apscheduler 
telethon-tgcrypto
mutagen
# yt-dlp removed - bot now only supports Telegram content downloads
speedtest-cli
python-dotenv
